// src/types/mock-server-data-types.ts
import { City, HousingType } from './offer';

export interface MockData {
    titles: string[];
    descriptions: string[];
    cities: City[];
    images: string[];
    types: HousingType[];
    goods: string[];
    names: string[];
    emails: string[];
    avatarPaths: string[];
    passwords: string[];
}